+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/IconSelect.kt
@@ -62,7 +62,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
             iconDrawableId = iconDrawableId,
             iconColorId = iconColorId,
             buttonMethod = buttonMethod,
-            view = requireView(),
+            view = requireView()
         )
     }
 
@@ -76,7 +76,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
             targetId = R.id.EditButton,
             iconDrawableId = R.drawable.arrow_pointer_icon_large,
             iconColorId = R.color.icon_primary,
-            buttonMethod = { viewModel.switchMode(Mode.EDIT) },
+            buttonMethod = { viewModel.switchMode(Mode.EDIT) }
         )
     }
 
@@ -84,7 +84,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
         convertButton(
             targetId = R.id.BackButton,
             iconDrawableId = R.drawable.arrow_left_icon_large,
-            buttonMethod = ::navigateToNewQuest,
+            buttonMethod = ::navigateToNewQuest
         )
     }
 
@@ -97,7 +97,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
         convertButton(
             targetId = R.id.AddNewIconButton,
             iconDrawableId = R.drawable.plus_icon_large,
-            buttonMethod = ::navigateToNewIcon,
+            buttonMethod = ::navigateToNewIcon
         )
     }
 
@@ -114,7 +114,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
             targetId = R.id.EditButton,
             iconDrawableId = R.drawable.pencil_icon_large,
             iconColorId = R.color.icon_primary,
-            buttonMethod = { viewModel.switchMode(Mode.DEFAULT) },
+            buttonMethod = { viewModel.switchMode(Mode.DEFAULT) }
         )
     }
 
@@ -169,7 +169,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
             targetId = R.id.BackButton,
             iconDrawableId = R.drawable.trash_bin_icon_large,
             iconColorId = R.color.warning_icon,
-            buttonMethod = ::deleteIcons,
+            buttonMethod = ::deleteIcons
         )
     }
 
@@ -198,7 +198,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
         convertButton(
             targetId = R.id.AddNewIconButton,
             iconDrawableId = R.drawable.dash_icon_large,
-            buttonMethod = ::setMoveMode,
+            buttonMethod = ::setMoveMode
         )
     }
 
@@ -411,7 +411,7 @@ class IconSelect : Fragment(R.layout.icon_select) {
         convertButton(
             targetId = R.id.AddNewIconButton,
             iconDrawableId = R.drawable.cancel_icon_large,
-            buttonMethod = ::cancelMove,
+            buttonMethod = ::cancelMove
         )
     }