+++ app/src/main/java/com/brandonjamesyoung/levelup/viewmodels/ShopViewModel.kt
@@ -2,10 +2,13 @@ package com.brandonjamesyoung.levelup.viewmodels
 
 import android.util.Log
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import com.brandonjamesyoung.levelup.constants.MAX_POINTS_PER_PURCHASE
 import com.brandonjamesyoung.levelup.constants.Mode
+import com.brandonjamesyoung.levelup.constants.SortOrder
+import com.brandonjamesyoung.levelup.constants.SortType
 import com.brandonjamesyoung.levelup.data.*
 import com.brandonjamesyoung.levelup.di.IoDispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -24,6 +27,23 @@ class ShopViewModel @Inject constructor(
     val shopItemList: LiveData<List<ShopItem>> = shopItemRepository.observeAll().asLiveData()
 
     val player: LiveData<Player> = playerRepository.observe().asLiveData()
+
+    private var _sortType: MutableLiveData<SortType> =
+        MutableLiveData<SortType>(SortType.DATE_CREATED)
+
+    val sortType: MutableLiveData<SortType>
+        get() = _sortType
+
+    private var _sortOrder: MutableLiveData<SortOrder> = MutableLiveData<SortOrder>(SortOrder.DESC)
+
+    val sortOrder: MutableLiveData<SortOrder>
+        get() = _sortOrder
+
+    private var validSortTypes: List<SortType> = listOf(
+        SortType.DATE_CREATED,
+        SortType.NAME,
+        SortType.PRICE
+    )
 
     init {
         validModes = listOf(Mode.DEFAULT, Mode.SELECT)
@@ -79,6 +99,20 @@ class ShopViewModel @Inject constructor(
         shopItemRepository.delete(ids)
         val numDeleted = ids.count()
         Log.i(TAG, "Delete $numDeleted item(s)")
+    }
+
+    fun switchSortType() {
+        if (_sortOrder.value == SortOrder.DESC) {
+            _sortOrder.value = SortOrder.ASC
+            return
+        }
+
+        // Go down the sort list and if end is reached loop back to beginning sort type
+        var currSortTypeIndex: Int = validSortTypes.indexOf(_sortType.value)
+        currSortTypeIndex += 1
+        if (currSortTypeIndex >= validSortTypes.size) currSortTypeIndex = 0
+        _sortType.value = validSortTypes[currSortTypeIndex]
+        _sortOrder.value = SortOrder.DESC
     }
 
     companion object {